refactor: zipXlfFiles becomes zipFiles in FileFunctions
