+++ extension/src/Common.ts
@@ -47,37 +47,6 @@ export function deleteFolderRecursive(directoryPath: string): void {
   }
 }
 
-export function mkDirByPathSync(targetDir: string): string {
-  const sep = path.sep;
-  const initDir = path.isAbsolute(targetDir) ? sep : "";
-  const baseDir = ".";
-
-  return targetDir.split(sep).reduce((parentDir, childDir) => {
-    const curDir = path.resolve(baseDir, parentDir, childDir);
-    try {
-      fs.mkdirSync(curDir);
-    } catch (err) {
-      if (err.code === "EEXIST") {
-        // curDir already exists!
-        return curDir;
-      }
-
-      // To avoid `EISDIR` error on Mac and `EACCES`-->`ENOENT` and `EPERM` on Windows.
-      if (err.code === "ENOENT") {
-        // Throw the original parentDir error on curDir `ENOENT` failure.
-        throw new Error(`EACCES: permission denied, mkdir '${parentDir}'`);
-      }
-
-      const caughtErr = ["EACCES", "EPERM", "EISDIR"].indexOf(err.code) > -1;
-      if (!caughtErr || (caughtErr && curDir === path.resolve(targetDir))) {
-        throw err; // Throw if it's just the last created dir.
-      }
-    }
-
-    return curDir;
-  }, initDir);
-}
-
 /**
  *
  * @param date Default Today
@@ -92,10 +61,4 @@ export function formatDate(date = new Date()): string {
   day = day.length < 2 ? `0${day}` : day;
 
   return `${year}-${month}-${day}`;
-}
-
-export function createFolderIfNotExist(folderPath: string): void {
-  if (!fs.existsSync(folderPath)) {
-    mkDirByPathSync(folderPath);
-  }
 }
