+++ extension/src/test/Common.test.ts
@@ -6,7 +6,6 @@ import * as Common from "../Common";
 suite("Common", function () {
   const parentPath = path.resolve(__dirname, "common-test");
   const newPath = path.resolve(parentPath, "new/path/", Common.formatDate());
-  const invalidPathChars = `/?<>\\:*|."`;
 
   test("convertLinefeedToBr", function () {
     assert.strictEqual(
@@ -29,33 +28,6 @@ suite("Common", function () {
       "2001-01-01",
       "Incorrect date string returned"
     );
-  });
-
-  test("createFolderIfNotExist", function () {
-    assert.strictEqual(
-      fs.existsSync(newPath),
-      false,
-      "Test folder should not exist. This could be an indication that the 'deleteFolderRecursive' test is not working."
-    );
-    Common.createFolderIfNotExist(newPath);
-    assert.ok(fs.existsSync(newPath), "Could not find created folder.");
-
-    // Test error code path
-    if (process.platform === "win32") {
-      // Linux: pretty much allows any character in a path.
-      // MacOS: We're currently not running test on MacOS.
-      let errorMsg = "";
-      try {
-        Common.createFolderIfNotExist(newPath + invalidPathChars);
-      } catch (e) {
-        errorMsg = (e as Error).message;
-      }
-      assert.strictEqual(
-        errorMsg.length > 0,
-        true,
-        "Expected Error to be thrown"
-      );
-    }
   });
 
   test("deleteFolderRecursive", function () {
